keyserver/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
cryptography==41.0.4
PyJWT==2.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
